functions: improve handling during idle moments
Before, one internal clock would signal a timeout from which a check
for load and hot containers would decide whether a new hot container
would have to be started.

This commit improves the internal clockwork by removing this internal
clock. Now, for each incoming task, a quick saturation check is done,
and if necessary a new hot container is started.